+++ Sources/StreamChatUI/Appearance+Images.swift
@@ -81,6 +81,7 @@ public extension Appearance {
 
         // MARK: - Polls
 
+        public var pollWinner: UIImage = loadSafely(systemName: "trophy", assetsFallback: "trophy")
         public var pollVoteCheckmarkActive: UIImage = .checkmark
         public var pollVoteCheckmarkInactive: UIImage = UIImage(
             systemName: "circle",